moved history moves into its own component
