+++ src/components/Dashboard/InGameStats/HistoryItem/HistoryItem.css
@@ -1,0 +1,22 @@
+.moves-line{
+    color: #FFFFFF;
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    
+    width: 100%;
+    height: 4.5vh;
+}
+
+.moves-scroll{
+    width: 105%;
+}
+
+.moves-line h3{
+    font-size: 14px;
+    flex: 1;
+}
+
+.move-number{
+    padding-left: 15%;
+}
